Moved fixXreacs.py from python/rdesigneur directory to python directory
